изменен тип переменных

+++ PP/Form1.cs
@@ -19,9 +19,9 @@ namespace PP
 
         private void button1_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToSingle(textBox1.Text);
+            double b = Convert.ToSingle(textBox2.Text);
+            double c;
             c = a + b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Сложение";
@@ -30,9 +30,9 @@ namespace PP
 
         private void button2_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToSingle(textBox1.Text);
+            double b = Convert.ToSingle(textBox2.Text);
+            double c;
             c = a - b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Вычитание";
@@ -41,9 +41,9 @@ namespace PP
 
         private void button3_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToSingle(textBox1.Text);
+            double b = Convert.ToSingle(textBox2.Text);
+            double c;
             c = a * b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Умножение";
@@ -52,9 +52,9 @@ namespace PP
 
         private void button4_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToSingle(textBox1.Text);
+            double b = Convert.ToSingle(textBox2.Text);
+            double c;
             c = a / b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Деление";
